# api/openapi.yaml
openapi: 3.1.0
info:
  title: Contextarium API
  version: 0.1.0-draft
  description: >
    Draft endpoint inventory plus already-frozen shared constraints for Contextarium.
    During M0-M6 this is not yet the complete production REST contract.
    Before an endpoint is implemented, its owning milestone must define its request,
    response, security, concurrency/idempotency, and error semantics. M7 reconciles
    and freezes REST v1. Implemented M2 record, revision, restore and minimal audit
    behavior follows docs/m2-contract.md; subjects/schemas retain docs/m1-contract.md.
    The runtime remains unauthenticated, loopback-only development software.
    Security declarations below describe the later v1 target; consult the temporary
    M2 contract for actual base_revision, pagination and legacy replay semantics.
...
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /subjects:
    get:
      summary: List authorized subjects
      operationId: listSubjects
      responses:
        "200":
          description: Subject list
    post:
      summary: Create subject
      operationId: createSubject
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Subject created
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /records:
    get:
      summary: List or filter authorized records
      operationId: listRecords
      parameters:
        - in: query
          name: namespace
          schema:
            type: string
        - in: query
          name: subject_id
          schema:
            type: string
      responses:
        "200":
          description: Record list
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      summary: Create an authoritative record
      operationId: createRecord
      description: Requires explicit direct-write authority. Agent clients should normally use proposals.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Record created
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Identity or create-precondition conflict
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Client bearer credential. Authenticated traffic over an untrusted network
        must use encrypted transport.
  parameters:
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
      description: >
        Required for retry-sensitive mutations. Exact storage/replay semantics are
        frozen by the milestone that implements the operation.
  responses:
    Unauthorized:
      description: Missing or invalid authentication
    Forbidden:
      description: Authenticated client is not authorized for the operation/resource
...